+++ src/main/java/de/ellpeck/actuallyadditions/mod/blocks/AABlockItem.java
@@ -1,12 +1,11 @@
 package de.ellpeck.actuallyadditions.mod.blocks;
 
-import net.minecraft.core.component.DataComponents;
+import de.ellpeck.actuallyadditions.mod.components.ActuallyComponents;
 import net.minecraft.network.chat.Component;
 import net.minecraft.world.item.BlockItem;
 import net.minecraft.world.item.ItemNameBlockItem;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.TooltipFlag;
-import net.minecraft.world.item.component.CustomData;
 import net.minecraft.world.level.block.Block;
 
 import javax.annotation.Nullable;
@@ -38,12 +37,8 @@ public class AABlockItem extends BlockItem {
         public void appendHoverText(ItemStack pStack, @Nullable TooltipContext pContext, List<Component> pTooltip, TooltipFlag pFlag) {
             super.appendHoverText(pStack, pContext, pTooltip, pFlag);
 
-            if (pStack.has(DataComponents.BLOCK_ENTITY_DATA)) {
-                CustomData customData = pStack.get(DataComponents.BLOCK_ENTITY_DATA);
-                int energy = 0;
-                if (customData.contains("Energy")) {
-                    energy = customData.copyTag().getInt("Energy");
-                }
+            if (pStack.has(ActuallyComponents.ENERGY_STORAGE)) {
+                int energy = pStack.getOrDefault(ActuallyComponents.ENERGY_STORAGE, 0);
                 NumberFormat format = NumberFormat.getInstance();
                 pTooltip.add(Component.translatable("misc.actuallyadditions.power_single", format.format(energy)));
             }